requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
numpy>=1.26.0
Pillow==10.0.1
scikit-learn>=1.3.0
pandas>=2.0.3
matplotlib>=3.7.2
seaborn>=0.12.2
tensorflow>=2.13.0
keras>=2.13.1
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
PyPDF2==3.0.1
pytesseract==0.3.10 